requirements.txt
asgiref==3.5.2
Django==3.2.14
django-jazzmin==2.5.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
